requirements.txt
crewai
python-dotenv
litellm
fastapi
uvicorn
pydantic
reportlab
